env: split the Wasmtime host into brain-env-worker
brain-server depends on brain-env for the Environment adapter and worker
supervision, but the crate also carried the Wasmtime engine, so the server's
dependency graph claimed a Wasm runtime it never calls. Move admission,
guest execution and the worker service into brain-env-worker, which owns the
binary of the same name and is now the only crate that links an engine.

brain-env keeps the client, supervisor, socket, wire and limits. NativeToolInput
moves to wire.rs as the only type that crossed the cut, and the frame codec and
limit accessors become public so the worker can reach them. The hand-written WIT
stays under crates/brain-env/wit: docs, fixtures, examples and the SDK generator
reference that path, and the worker's bindgen reads it relatively.

No behaviour changes. Same two processes, same spawn, same wire.
