+++ .github/workflows/ci.yml
@@ -96,7 +96,7 @@ jobs:
           BRAIN_TEST_AGENTLOOP_PACKAGE: ${{ github.workspace }}/tests/fixtures/diagnostic-agentloop/target/wasm32-wasip2/release/diagnostic_agentloop.wasm
           BRAIN_TEST_TOOL_COMPONENT: ${{ github.workspace }}/tests/fixtures/diagnostic-tool/target/wasm32-wasip2/release/diagnostic_tool.wasm
           BRAIN_TEST_REFERENCE_AGENTLOOP: ${{ github.workspace }}/examples/reference-agentloop/target/wasm32-wasip2/release/reference_agentloop.wasm
-        run: cargo test -p brain-env --features integration-tests --test worker_process
+        run: cargo test -p brain-env-worker --features integration-tests --test worker_process
 
   brain-env:
     runs-on: ubuntu-24.04
@@ -127,14 +127,14 @@ jobs:
           BRAIN_TEST_AGENTLOOP_PACKAGE: ${{ runner.temp }}/diagnostic-agentloop.wasm
           BRAIN_TEST_TOOL_COMPONENT: ${{ runner.temp }}/diagnostic-tool.wasm
           BRAIN_TEST_REFERENCE_AGENTLOOP: ${{ runner.temp }}/reference-agentloop.wasm
-        run: cargo test -p brain-env --features integration-tests --test worker_process --test in_process
+        run: cargo test -p brain-env-worker --features integration-tests --test worker_process --test in_process
 
       - uses: actions/setup-node@v7
         with:
           node-version: "22.23.2"
           cache: npm
       - run: npm ci && npm run build
-      - run: cargo build -p brain-server --bin brain -p brain-env --bin brain-env-worker
+      - run: cargo build -p brain-server --bin brain -p brain-env-worker --bin brain-env-worker
       - name: Exercise lazy providers, cold history, and retained sessions through HTTP
         env:
           BRAIN_TEST_SERVER: ${{ github.workspace }}/target/debug/brain
