+++ crates/brain-env-worker/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "brain-env-worker"
+version = "0.1.0"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+description = "Bounded Wasmtime Component host for Brain Agentloops"
+publish = false
+autobins = false
+
+[features]
+integration-tests = []
+
+[[bin]]
+name = "brain-env-worker"
+path = "src/main.rs"
+
+[[test]]
+name = "worker_process"
+path = "tests/worker_process.rs"
+required-features = ["integration-tests"]
+
+[[test]]
+name = "in_process"
+path = "tests/in_process.rs"
+required-features = ["integration-tests"]
+
+[dependencies]
+brain-env = { workspace = true }
+brain-protocol = { workspace = true }
+anyhow = { workspace = true }
+async-trait = { workspace = true }
+base64 = { workspace = true }
+clap = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+sha2 = { workspace = true }
+http = { workspace = true }
+http-body-util = { workspace = true }
+tokio = { workspace = true, features = ["fs", "process", "net", "io-util", "sync", "macros", "rt-multi-thread"] }
+url = { workspace = true }
+wasmtime = { workspace = true }
+wasmtime-wasi = { workspace = true }
+wasmtime-wasi-http = { workspace = true }
+tempfile = { workspace = true }
+
+[dev-dependencies]
+# Builds the smallest possible runaway guest for the fuel-exhaustion test. Already in the
+# lock file as a Wasmtime component dependency, so it costs no new crate.
+wasm-encoder = "0.254"
