+++ crates/brain-env/src/wire.rs
@@ -145,7 +145,15 @@ pub enum WorkerResponse {
     },
 }
 
-pub(crate) async fn write_frame<W: AsyncWrite + Unpin, T: Serialize>(
+/// What a native Tool Component is handed for one call.
+#[derive(serde::Serialize, serde::Deserialize)]
+pub struct NativeToolInput {
+    pub input: serde_json::Value,
+    pub configuration: serde_json::Value,
+    pub deadline_at_ms: u64,
+}
+
+pub async fn write_frame<W: AsyncWrite + Unpin, T: Serialize>(
     writer: &mut W,
     value: &T,
     max_bytes: usize,
@@ -167,7 +175,7 @@ pub(crate) async fn write_frame<W: AsyncWrite + Unpin, T: Serialize>(
     writer.flush().await.map_err(|error| error.to_string())
 }
 
-pub(crate) async fn read_frame<R: AsyncRead + Unpin, T: for<'de> Deserialize<'de>>(
+pub async fn read_frame<R: AsyncRead + Unpin, T: for<'de> Deserialize<'de>>(
     reader: &mut R,
     max_bytes: usize,
 ) -> Result<T, String> {
@@ -183,7 +191,7 @@ pub(crate) async fn read_frame<R: AsyncRead + Unpin, T: for<'de> Deserialize<'de
     serde_json::from_slice(&payload).map_err(|error| error.to_string())
 }
 
-pub(crate) fn max_request_bytes(request: &WorkerRequest, limits: &EnvLimits) -> usize {
+pub fn max_request_bytes(request: &WorkerRequest, limits: &EnvLimits) -> usize {
     match request {
         WorkerRequest::Ping | WorkerRequest::Cancel => 1_024,
         WorkerRequest::Admit { .. } => limits.max_request_frame_bytes(),
