+++ tests/journeys/README.md
@@ -33,7 +33,7 @@ tests synchronize with actual model/Tool entry rather than guessing execution pr
 After `npm ci && npm run build`:
 
 ```sh
-cargo build -p brain-server --bin brain -p brain-env --bin brain-env-worker
+cargo build -p brain-server --bin brain -p brain-env-worker --bin brain-env-worker
 cargo build --manifest-path tests/fixtures/diagnostic-agentloop/Cargo.toml --target wasm32-wasip2 --release
 cargo build --manifest-path tests/fixtures/diagnostic-tool/Cargo.toml --target wasm32-wasip2 --release
 cargo build --manifest-path examples/reference-agentloop/Cargo.toml --target wasm32-wasip2 --release
